feat(playback): Tier-2 NDI auto-recovery + drop dashboard alert
Three changes flipping the NDI work from "observe" to "fix":

1. Drop the NdiHealthCard dashboard component. The operator wanted a
   working app, not a fancy alert about a broken one. Backend telemetry
   stays (registry + /api/v1/ndi/health endpoint for forensics + tests).

2. Tighten the degradation gate to require scene_active=true. A pipeline
   that's still in PlayState::Playing while its OBS scene is off-program
   is normal mid-transition; reporting connections=0 there is noise.
   Now we only mark a pipeline degraded when OBS expects this source
   AND isn't getting frames.

3. Tier-2 auto-recovery: on every 30 s (6th consecutive bad poll) of
   sustained degradation, the engine sends PipelineCommand::RecreateSender
   to the pipeline thread. The thread destroys the existing NDI sender
   (Drop calls send_video_flush + send_destroy) and creates a fresh one
   with the same name, which forces OBS receivers to re-handshake.
   Handles both the outer wait loop and the inner active-decode loop.

This actually fixes the #46 dark-wall-after-restart symptom: the system
self-heals within ~30 s instead of requiring an operator scene-flip.
