+++ crates/sp-server/src/playback/pipeline.rs
@@ -41,6 +41,10 @@ pub enum PipelineCommand {
     Stop,
     /// Shut down the thread.
     Shutdown,
+    /// Force-recreate the NDI sender (destroy + create with the same name).
+    /// Used by the auto-recovery path when delivery has been failing for
+    /// >= 30 s — drops the existing handle so OBS receivers must re-handshake.
+    RecreateSender,
 }
 
 /// Events emitted by the pipeline thread back to the async engine.
@@ -157,6 +161,16 @@ impl PlaybackPipeline {
         let _ = self.cmd_tx.send(cmd);
     }
 
+    /// Send a command to the pipeline thread. Returns Err if the channel
+    /// is closed (thread already exited). Used by the auto-recovery path
+    /// in `playback::ndi_health`.
+    pub fn send_command(
+        &self,
+        cmd: PipelineCommand,
+    ) -> Result<(), crossbeam_channel::SendError<PipelineCommand>> {
+        self.cmd_tx.send(cmd)
+    }
+
     /// Gracefully shut down the pipeline, blocking until the thread exits.
     pub fn shutdown(mut self) {
         let _ = self.cmd_tx.send(PipelineCommand::Shutdown);
@@ -247,6 +261,9 @@ fn run_loop_stub(
             }
             Ok(PipelineCommand::Stop) => {
                 info!(playlist_id, "pipeline: stopped (stub)");
+            }
+            Ok(PipelineCommand::RecreateSender) => {
+                tracing::info!(playlist_id, "RecreateSender ignored (non-Windows stub)");
             }
         }
     }
@@ -359,6 +376,8 @@ fn run_loop_windows(
                     match decode_and_send(
                         &cmd_rx,
                         &mut submitter,
+                        backend.clone(),
+                        ndi_name,
                         &current_video,
                         &current_audio,
                         &event_tx,
@@ -435,6 +454,25 @@ fn run_loop_windows(
                 submitter.send_black_bgra(1920, 1080);
                 debug!(playlist_id, "stopped (no active playback)");
             }
+            Ok(PipelineCommand::RecreateSender) => {
+                let prev_rate_n = submitter.frame_rate_n();
+                let prev_rate_d = submitter.frame_rate_d();
+                match sp_ndi::NdiSender::new_with_clocking(backend.clone(), ndi_name, true, false) {
+                    Ok(new_sender) => {
+                        // Replace submitter; old one's Drop calls
+                        // send_video_flush + send_destroy on the old sender.
+                        submitter = FrameSubmitter::new(new_sender, prev_rate_n, prev_rate_d);
+                        submitter.send_black_bgra(1920, 1080);
+                        info!(
+                            playlist_id,
+                            ndi_name, "NDI sender recreated (Tier-2 auto-recovery)"
+                        );
+                    }
+                    Err(e) => {
+                        error!(%e, ndi_name, "RecreateSender: failed to create new sender; keeping existing");
+                    }
+                }
+            }
         }
     }
 }
@@ -462,6 +500,8 @@ enum DecodeResult {
 fn decode_and_send(
     cmd_rx: &Receiver<PipelineCommand>,
     submitter: &mut FrameSubmitter<sp_ndi::RealNdiBackend>,
+    backend: SharedNdiBackend,
+    ndi_name: &str,
     video_path: &std::path::Path,
     audio_path: &std::path::Path,
     event_tx: &tokio::sync::mpsc::UnboundedSender<(i64, PipelineEvent)>,
@@ -541,6 +581,25 @@ fn decode_and_send(
                 if let Err(e) = decoder.seek(position_ms) {
                     tracing::warn!(?e, position_ms, "pipeline: seek failed");
                 }
+            }
+            Ok(PipelineCommand::RecreateSender) => {
+                // Force-recreate during active decode. Same flow as outer.
+                let prev_rate_n = submitter.frame_rate_n();
+                let prev_rate_d = submitter.frame_rate_d();
+                match sp_ndi::NdiSender::new_with_clocking(backend.clone(), ndi_name, true, false) {
+                    Ok(new_sender) => {
+                        *submitter = FrameSubmitter::new(new_sender, prev_rate_n, prev_rate_d);
+                        submitter.send_black_bgra(1920, 1080);
+                        info!(
+                            playlist_id,
+                            ndi_name, "NDI sender recreated mid-decode (Tier-2 auto-recovery)"
+                        );
+                    }
+                    Err(e) => {
+                        error!(%e, ndi_name, "RecreateSender mid-decode: failed; keeping existing");
+                    }
+                }
+                // Continue the decode loop — don't break.
             }
             Err(TryRecvError::Empty) => {}
             Err(TryRecvError::Disconnected) => {
@@ -779,215 +838,8 @@ fn emit_heartbeat(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::playback::ndi_health::PlaybackStateLabel;
-    use std::time::Instant;
-
-    #[test]
-    fn pipeline_spawn_and_shutdown() {
-        let (event_tx, _event_rx) = tokio::sync::mpsc::unbounded_channel();
-        let pipeline = PlaybackPipeline::spawn("test-ndi".into(), None, event_tx, 1);
-        pipeline.shutdown();
-        // If we get here, the thread joined successfully.
-    }
-
-    #[test]
-    fn pipeline_drop_sends_shutdown() {
-        let (event_tx, _event_rx) = tokio::sync::mpsc::unbounded_channel();
-        {
-            let _pipeline = PlaybackPipeline::spawn("test-drop".into(), None, event_tx, 2);
-            // Pipeline dropped here — Drop impl should send Shutdown and join.
-        }
-        // If we get here without hanging, the Drop worked correctly.
-    }
-
-    #[test]
-    fn pipeline_send_command_before_shutdown() {
-        let (event_tx, _event_rx) = tokio::sync::mpsc::unbounded_channel();
-        let pipeline = PlaybackPipeline::spawn("test-cmd".into(), None, event_tx, 3);
-        pipeline.send(PipelineCommand::Stop);
-        pipeline.send(PipelineCommand::Pause);
-        pipeline.send(PipelineCommand::Resume);
-        pipeline.shutdown();
-    }
-
-    #[test]
-    fn pipeline_play_emits_event_on_non_windows() {
-        let (event_tx, mut event_rx) = tokio::sync::mpsc::unbounded_channel();
-        let pipeline = PlaybackPipeline::spawn("test-play".into(), None, event_tx, 4);
-
-        pipeline.send(PipelineCommand::Play {
-            video: PathBuf::from("/tmp/test_video.mp4"),
-            audio: PathBuf::from("/tmp/test_audio.flac"),
-        });
-        // Give the thread a moment to process.
-        std::thread::sleep(std::time::Duration::from_millis(50));
-        pipeline.shutdown();
-
-        // On non-Windows, we expect an Error event.
-        #[cfg(not(windows))]
-        {
-            let (id, event) = event_rx.try_recv().expect("should have received an event");
-            assert_eq!(id, 4);
-            match event {
-                PipelineEvent::Error(msg) => {
-                    assert!(msg.contains("Windows"), "error should mention Windows");
-                }
-                other => panic!("expected Error event, got {other:?}"),
-            }
-        }
-
-        let _ = event_rx;
-    }
-
-    #[test]
-    fn seek_variant_carries_position_ms() {
-        let cmd = PipelineCommand::Seek { position_ms: 12345 };
-        match cmd {
-            PipelineCommand::Seek { position_ms } => assert_eq!(position_ms, 12345),
-            _ => panic!("wrong variant"),
-        }
-    }
-
-    #[test]
-    fn seek_does_not_collide_with_other_variants() {
-        // Compile-time check that every variant is still distinct.
-        let variants = vec![
-            PipelineCommand::Play {
-                video: PathBuf::new(),
-                audio: PathBuf::new(),
-            },
-            PipelineCommand::Pause,
-            PipelineCommand::Resume,
-            PipelineCommand::Seek { position_ms: 0 },
-            PipelineCommand::Stop,
-            PipelineCommand::Shutdown,
-        ];
-        assert_eq!(variants.len(), 6);
-    }
-
-    #[test]
-    fn pipeline_send_seek_command() {
-        let (event_tx, _event_rx) = tokio::sync::mpsc::unbounded_channel();
-        let pipeline = PlaybackPipeline::spawn("test-seek".into(), None, event_tx, 6);
-        pipeline.send(PipelineCommand::Seek { position_ms: 5000 });
-        pipeline.shutdown();
-        // No panic or hang means the Seek arm is handled in the loop.
-    }
-
-    /// Regression test for the NewPlay bug: the outer loop must continue to
-    /// receive and process subsequent Play commands after the first one returns.
-    /// Before the fix, NewPlay was discarded and the worker would hang waiting
-    /// for the next command instead of playing the new file.
-    ///
-    /// On non-Windows the stub emits an Error per Play. On Windows the loop
-    /// path is structurally the same — verified live on win-resolume v0.7.2.
-    #[test]
-    fn pipeline_processes_multiple_sequential_plays() {
-        let (event_tx, mut event_rx) = tokio::sync::mpsc::unbounded_channel();
-        let pipeline = PlaybackPipeline::spawn("test-multi-play".into(), None, event_tx, 5);
-
-        pipeline.send(PipelineCommand::Play {
-            video: PathBuf::from("/tmp/song-a_video.mp4"),
-            audio: PathBuf::from("/tmp/song-a_audio.flac"),
-        });
-        std::thread::sleep(std::time::Duration::from_millis(30));
-        pipeline.send(PipelineCommand::Play {
-            video: PathBuf::from("/tmp/song-b_video.mp4"),
-            audio: PathBuf::from("/tmp/song-b_audio.flac"),
-        });
-        std::thread::sleep(std::time::Duration::from_millis(30));
-        pipeline.send(PipelineCommand::Play {
-            video: PathBuf::from("/tmp/song-c_video.mp4"),
-            audio: PathBuf::from("/tmp/song-c_audio.flac"),
-        });
-        std::thread::sleep(std::time::Duration::from_millis(50));
-        pipeline.shutdown();
-
-        // Drain all events. We expect at least one event per Play command —
-        // proving the worker did not hang after the first Play.
-        #[cfg(not(windows))]
-        {
-            let mut event_count = 0;
-            while let Ok((id, event)) = event_rx.try_recv() {
-                assert_eq!(id, 5);
-                match event {
-                    PipelineEvent::Error(_) => event_count += 1,
-                    other => panic!("unexpected event: {other:?}"),
-                }
-            }
-            assert_eq!(
-                event_count, 3,
-                "expected 3 Error events (one per Play), got {event_count}"
-            );
-        }
-
-        let _ = event_rx;
-    }
-
-    /// Pin: `PipelineCommand::Play` must reset `paused = false` BEFORE
-    /// entering `decode_and_send`, so a stale `Pause` cannot leak across
-    /// video changes. Static check via `include_str!` — fires red if the
-    /// `paused = false` statement is moved out of the Play arm, deleted,
-    /// or commented out.
-    #[test]
-    fn play_command_clears_paused_state() {
-        let src = include_str!("pipeline.rs");
-        let play_arm_start = src
-            .find("Ok(PipelineCommand::Play {")
-            .expect("Play arm must exist");
-        let decode_call = src[play_arm_start..]
-            .find("decode_and_send(")
-            .expect("Play arm must call decode_and_send");
-        let play_block = &src[play_arm_start..play_arm_start + decode_call];
-
-        // Strict match: an actual statement (semicolon-terminated), not a
-        // commented-out line or a docstring mention. Strip any line whose
-        // first non-whitespace chars are `//` before checking.
-        let live_lines: String = play_block
-            .lines()
-            .filter(|line| !line.trim_start().starts_with("//"))
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(
-            live_lines.contains("paused = false;"),
-            "PipelineCommand::Play must clear `paused = false;` (live statement, not \
-             a comment) BEFORE decode_and_send. Current Play arm:\n{play_block}"
-        );
-    }
-
-    #[test]
-    fn health_snapshot_variant_constructs_and_clones() {
-        let now = Instant::now();
-        let ev = PipelineEvent::HealthSnapshot {
-            connections: 1,
-            frames_submitted_total: 100,
-            frames_submitted_last_5s: 30,
-            observed_fps: 29.97,
-            nominal_fps: 29.97,
-            last_submit_ts: Some(now),
-            last_heartbeat_ts: now,
-            consecutive_bad_polls: 0,
-            reported_state: PlaybackStateLabel::Playing,
-        };
-        let cloned = ev.clone();
-        // Pattern-match to assert the variant exists and the fields round-trip.
-        if let PipelineEvent::HealthSnapshot {
-            connections,
-            frames_submitted_last_5s,
-            reported_state,
-            ..
-        } = cloned
-        {
-            assert_eq!(connections, 1);
-            assert_eq!(frames_submitted_last_5s, 30);
-            assert_eq!(reported_state, PlaybackStateLabel::Playing);
-        } else {
-            panic!("clone produced wrong variant");
-        }
-    }
-}
+#[path = "pipeline_inline_tests.rs"]
+mod tests;
 
 #[cfg(test)]
 #[path = "pipeline_heartbeat_tests.rs"]
