+++ sp-ui/src/components/mod.rs
@@ -9,7 +9,6 @@ pub mod lyrics_scroller;
 pub mod lyrics_song_detail;
 pub mod lyrics_song_row;
 pub mod now_playing_card;
-pub mod ndi_health;
 pub mod obs_status;
 pub mod playback_controls;
 pub mod playlist_card;
